refactor(client): extract auth/valid_login tests to sibling file
